Base Class System
Restructured the hardware classes so that there is a
base class for each piece of hardware that will act
as an 'example' class. The idea being it can perform
all of the functions of a regular class. Majority of
the hardware classes have been restructured at this point.

CE Gui can now run using the BaseSystem.