Safe Move Implementation
Have not tested Safe Move yet in class form, but did test in functional form and worked okay.

There are at most 3 moves that need to be done safely:

1. Raise the capillary up
2. Move the Stage
3. Lower the capillary

You cannot move the stage until the capillary won't crash into something (an increasing z requirement). And you cannot lower the capillary unless there are no remaining ledges the capillary will crash into. Time delays and kinematics are used to control and predict the location of the capillary and stage.